requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
PyYAML>=6.0
requests>=2.31.0
pytz>=2023.3
lxml>=4.9.0
